requirements.txt
pyautogui
pillow~=10.4.0
tkcap~=0.0.3
pydicom~=2.4.4
img2pdf
opencv_python
matplotlib
pandas
tensorflow==2.15.0
python-xlib

opencv-python~=4.10.0.84
numpy~=1.26.4